feat: add revert to registration

+++ AuthMicroservice/src/main/java/com/auth/saga/workflow/impl/ProfileWorkflowStep.java
@@ -16,6 +16,8 @@ public class ProfileWorkflowStep implements WorkflowStep {
     private ProfileResponseDTO profileResponseDTO;
     private WorkflowStepStatus stepStatus = WorkflowStepStatus.PENDING;
     private boolean started = false;
+    private Mono<Boolean> request;
+
     private final String uri = "api/v1/profiles";
 
     public ProfileWorkflowStep(WebClient webClient, RegisterDTO requestDTO) {
@@ -30,8 +32,8 @@ public class ProfileWorkflowStep implements WorkflowStep {
 
     @Override
     public Mono<Boolean> process() {
-        started = true;
-        return this.webClient
+        stepStatus = WorkflowStepStatus.START;
+        request = webClient
                 .post()
                 .uri(uri)
                 .header(HttpHeaders.CONTENT_TYPE, MediaType.APPLICATION_JSON_VALUE)
@@ -43,11 +45,15 @@ public class ProfileWorkflowStep implements WorkflowStep {
                 .map(ProfileResponseDTO::isSuccess)
                 //.map(r -> true)
                 .doOnNext(b -> this.stepStatus = b ? WorkflowStepStatus.COMPLETE : WorkflowStepStatus.FAILED);
+        return request;
     }
 
     @Override
     public Mono<Boolean> revert() {
-        while(this.stepStatus == WorkflowStepStatus.PENDING && started){}
+        request.block();
+        if (this.stepStatus == WorkflowStepStatus.FAILED)
+            return Mono.just(true);
+        //while(this.stepStatus == WorkflowStepStatus.PENDING && started){}
         return this.webClient
                 .delete()
                 .uri(uri + "/" + profileResponseDTO.getId())
